+++ kotlin-test/common/main/kotlin/com/javiersc/kotlin/test/fail.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.test
+
+import kotlin.test.asserter
+
+@PublishedApi internal fun fail(message: String?): Nothing = asserter.fail(message)