+++ kotlin-test/common/main/kotlin/com/javiersc/kotlin/test/iterableMatchers.kt
@@ -1,0 +1,23 @@
+package com.javiersc.kotlin.test
+
+public inline fun <reified I : Iterable<T>, reified T : Any> I.assertContains(
+    element: T,
+    message: String? = null,
+): I {
+    if (element in this) return this
+    fail(message ?: "Expected to contain $element but did not")
+}
+
+public inline fun <reified I : Iterable<T>, reified T : Any> I.assertCount(
+    expectSize: Int,
+    message: String? = null,
+): I = assertSize(expectSize = expectSize, message = message)
+
+public inline fun <reified I : Iterable<T>, reified T : Any> I.assertSize(
+    expectSize: Int,
+    message: String? = null,
+): I {
+    val actualSize: Int = count()
+    if (actualSize == expectSize) return this
+    fail(message ?: "Expected size $expectSize but was $actualSize")
+}